move rendering to dedicated file
